SLSCMN-3 organize tests
